1-ai_foundry_agent/promptflow/requirements.txt
aiohttp
azure-ai-inference
openai
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
azure-ai-ml
azure-ai-projects
azure-identity
azure-monitor-opentelemetry
bs4
readability
readability-lxml
lxml_html_clean